clean up copied files
