+++ components/src/AsyncSelect/AsyncSelect.js
@@ -1,13 +1,12 @@
 import React, { useContext } from "react";
 import AsyncReactSelect from "react-select/async";
-import { components as selectComponents } from "react-select";
 import { useTranslation } from "react-i18next";
 import { ThemeContext } from "styled-components";
+import PropTypes from "prop-types";
 import { Field } from "../Form";
 import { MaybeFieldLabel } from "../FieldLabel";
 import { InlineValidation } from "../Validation";
 import customStyles from "../Select/customReactSelectStyles";
-import { SelectPropTypes, SelectDefaultProps } from "./Select.type";
 import { SelectOption } from "./SelectOption";
 import { Control, MultiValue, ClearIndicator, SelectContainer, Menu, Input } from "../Select/SelectComponents";
 
@@ -107,8 +106,80 @@ const AsyncSelect = ({
   );
 };
 
-AsyncSelect.propTypes = SelectPropTypes;
+const valueType = PropTypes.oneOfType([
+  PropTypes.string,
+  PropTypes.number,
+  PropTypes.arrayOf(PropTypes.string),
+  PropTypes.arrayOf(PropTypes.number)
+]);
 
-AsyncSelect.defaultProps = SelectDefaultProps;
+AsyncSelect.propTypes = {
+  autocomplete: PropTypes.bool,
+  disabled: PropTypes.bool,
+  error: PropTypes.bool,
+  errorMessage: PropTypes.string,
+  errorList: PropTypes.arrayOf(PropTypes.string),
+  labelText: PropTypes.string,
+  helpText: PropTypes.node,
+  noOptionsMessage: PropTypes.func,
+  requirementText: PropTypes.string,
+  id: PropTypes.string,
+  initialIsOpen: PropTypes.bool,
+  menuPosition: PropTypes.string,
+  maxHeight: PropTypes.string,
+  multiselect: PropTypes.bool,
+  name: PropTypes.string,
+  onBlur: PropTypes.func,
+  onChange: PropTypes.func,
+  placeholder: PropTypes.string,
+  required: PropTypes.bool,
+  value: valueType,
+  defaultValue: valueType,
+  className: PropTypes.string,
+  classNamePrefix: PropTypes.string,
+  menuIsOpen: PropTypes.bool,
+  onMenuOpen: PropTypes.func,
+  onMenuClose: PropTypes.func,
+  onInputChange: PropTypes.func,
+  components: PropTypes.shape({}),
+  "aria-label": PropTypes.string,
+  cacheOptions: PropTypes.bool,
+  defaultOptions: PropTypes.arrayOf(PropTypes.shape({})),
+  loadOptions: PropTypes.func.isRequired
+};
+
+AsyncSelect.defaultProps = {
+  autocomplete: true,
+  disabled: null,
+  defaultValue: undefined,
+  error: undefined,
+  errorMessage: null,
+  errorList: null,
+  labelText: null,
+  helpText: null,
+  noOptionsMessage: () => null,
+  requirementText: null,
+  id: null,
+  initialIsOpen: undefined,
+  maxHeight: "248px",
+  menuPosition: "absolute",
+  multiselect: false,
+  name: undefined,
+  onBlur: undefined,
+  onChange: undefined,
+  placeholder: undefined,
+  required: false,
+  value: undefined,
+  className: null,
+  classNamePrefix: undefined,
+  menuIsOpen: undefined,
+  onMenuOpen: undefined,
+  onMenuClose: undefined,
+  onInputChange: undefined,
+  components: undefined,
+  "aria-label": undefined,
+  cacheOptions: false,
+  defaultOptions: undefined
+};
 
 export default AsyncSelect;
